requirements.txt
flask
httpx
pyyaml
bcrypt
cryptography
gunicorn
markdown
